+++ crates/agentgateway/src/llm/messages.rs
@@ -909,15 +909,7 @@ pub mod ingress {
         Ok(())
     }
 
-    // ===== HEADER EXTRACTION FUNCTIONS =====
-
     /// Extract Anthropic-specific headers for vendor storage
-    /// 
-    /// Returns a vendor data structure containing:
-    /// - anthropic-version: Required API version header
-    /// - anthropic-beta: Optional beta features (comma-separated or multiple headers)
-    /// 
-    /// Only records headers when provider is Anthropic for observability and forwarding.
     pub fn extract_anthropic_headers(headers: &HeaderMap) -> Option<Value> {
         let mut anthropic_headers = Map::new();
         
@@ -1553,140 +1545,6 @@ pub mod egress {
         })
     }
 
-    /// Wrap Universal SSE and re-emit Anthropic SSE in canonical order
-    /// 
-    /// Transforms Universal streaming to Anthropic format:
-    /// message_start → content_block_start → content_block_delta* → content_block_stop → message_delta* → message_stop
-    pub fn sse_universal_to_anthropic(body: Body) -> Body {
-        use crate::parse::sse;
-        
-        // For now, emit single events and ensure proper ordering
-        // TODO: Implement proper 1:N event emission
-        let mut first_chunk = true;
-        
-        sse::json_transform::<universal::StreamResponse, types::StreamEvent>(body, move |universal_chunk_result| {
-            match universal_chunk_result {
-                Ok(chunk) => {
-                    // Emit message_start for the first chunk
-                    if first_chunk && !chunk.choices.is_empty() {
-                        first_chunk = false;
-                        return Some(types::StreamEvent::MessageStart {
-                            message: types::MessagesResponse {
-                                id: chunk.id.clone(),
-                                r#type: "message".to_string(),
-                                role: "assistant".to_string(),
-                                model: chunk.model.clone(),
-                                content: Vec::new(),
-                                stop_reason: None,
-                                stop_sequence: None,
-                                usage: chunk.usage.map(|u| types::Usage {
-                                    input_tokens: u.prompt_tokens,
-                                    output_tokens: u.completion_tokens,
-                                    cache_creation_input_tokens: u.prompt_tokens_details.as_ref().and_then(|d| d.cached_tokens),
-                                    cache_read_input_tokens: u.prompt_tokens_details.as_ref().and_then(|d| d.cached_tokens),
-                                    cache_creation: None,
-                                    server_tool_use: None,
-                                    service_tier: None,
-                                }).unwrap_or(types::Usage {
-                                    input_tokens: 0,
-                                    output_tokens: 0,
-                                    cache_creation_input_tokens: None,
-                                    cache_read_input_tokens: None,
-                                    cache_creation: None,
-                                    server_tool_use: None,
-                                    service_tier: None,
-                                }),
-                                container: None,
-                            },
-                        });
-                    }
-                    
-                    // Transform the chunk - this will return multiple events but we can only emit one
-                    // FIXME: This is still dropping events but is better than the original bug
-                    if let Ok(events) = transform_universal_chunk_to_anthropic(chunk) {
-                        // For now, prioritize message_stop over other events to avoid dropping it
-                        for event in &events {
-                            if matches!(event, types::StreamEvent::MessageStop) {
-                                return Some(event.clone());
-                            }
-                        }
-                        // If no message_stop, return the first event
-                        events.into_iter().next()
-                    } else {
-                        None
-                    }
-                },
-                Err(_) => None,
-            }
-        })
-    }
-    
-    /// Transform a Universal stream chunk to Anthropic Messages API stream event
-    fn transform_universal_chunk_to_anthropic(chunk: universal::StreamResponse) -> Result<Vec<types::StreamEvent>, AIError> {
-        let mut events = Vec::new();
-        
-        // Get the first choice (assuming single choice for now)
-        let choice = chunk.choices.first()
-            .ok_or_else(|| AIError::MessageNotFound)?;
-        
-        // Handle different types of content
-        if let Some(content) = &choice.delta.content {
-            // Text content delta
-            events.push(types::StreamEvent::ContentBlockDelta {
-                index: choice.index as usize,
-                delta: types::ContentDelta::TextDelta { 
-                    text: content.clone() 
-                },
-            });
-        }
-        
-        // Handle tool calls if present
-        if let Some(tool_calls) = &choice.delta.tool_calls {
-            for tool_call in tool_calls {
-                if let Some(function) = &tool_call.function {
-                    if let Some(args) = &function.arguments {
-                        events.push(types::StreamEvent::ContentBlockDelta {
-                            index: choice.index as usize,
-                            delta: types::ContentDelta::InputJsonDelta { 
-                                partial_json: args.clone() 
-                            },
-                        });
-                    }
-                }
-            }
-        }
-        
-        // Handle completion (message_stop)
-        if let Some(finish_reason) = &choice.finish_reason {
-            let stop_reason = match finish_reason {
-                universal::FinishReason::Stop => Some(types::StopReason::EndTurn),
-                universal::FinishReason::Length => Some(types::StopReason::MaxTokens),
-                universal::FinishReason::ToolCalls => Some(types::StopReason::ToolUse),
-                universal::FinishReason::ContentFilter => Some(types::StopReason::Refusal),
-                _ => Some(types::StopReason::EndTurn),
-            };
-            
-            events.push(types::StreamEvent::MessageDelta {
-                delta: types::MessageDelta {
-                    stop_reason,
-                    stop_sequence: None,
-                    usage: chunk.usage.map(|u| types::Usage {
-                        input_tokens: u.prompt_tokens,
-                        output_tokens: u.completion_tokens,
-                        cache_creation_input_tokens: None,
-                        cache_read_input_tokens: None,
-                        cache_creation: None,
-                        server_tool_use: None,
-                        service_tier: None,
-                    }),
-                },
-            });
-            
-            events.push(types::StreamEvent::MessageStop);
-        }
-        
-        Ok(events)
-    }
 }
 
 // Re-export commonly used types for convenience
@@ -1697,7 +1555,7 @@ pub use types::{
 };
 
 pub use ingress::{to_universal, validate_messages_request};
-pub use egress::{from_universal_json, sse_universal_to_anthropic};
+pub use egress::from_universal_json;
 
 /// Fast token estimator for Messages API requests (optional optimization)
 pub fn estimate_tokens(request: &MessagesRequest) -> u32 {
